+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 ## 0.101.0 (Unreleased)
 
+FEATURES:
+* compute: added support for `instance_tags_pool` in `yandex_compute_instance_group`
+* compute: added support for `ignore_health_checks` in `yandex_compute_instance_group`
+
+
 BUG FIXES:
 * kafka: fix bug with update of field `schema_registry` in `yandex_mdb_kafka_cluster` resource
+* compute: fix `filesystem` in `compute_instance_group` resource
 
 ## 0.100.0 (October 11, 2023)